==> 01_UKBB_data_extraction/00_extract_UKBB_metadata.ipynb <==
### Packages

# pip install webdriver-manager

# pip install selenium
# pip install webdriver-manager

# Packages for webscraping
from selenium import webdriver
from selenium.webdriver.chrome.service import Service
from webdriver_manager.chrome import ChromeDriverManager
from selenium.common.exceptions import NoSuchElementException
from selenium.webdriver.common.by import By

#Package to handle dataframes
import pandas as pd

#Package for progress bar
from tqdm import tqdm

#### Configure webdriver to use Chrome browser

s=Service(ChromeDriverManager().install())
driver = webdriver.Chrome(service=s)

### Webscraping function

def UKBB_to_DF(ukbbidlist):
    # Initiate empty lists to store data of interest
    Description=[]
    Category = []
    ValueType=[]
    Nbpart = []
    Sex=[] 
    Mean=[]
    Median=[]
    Stdv=[]

    # 
    for ukbbid in tqdm(ukbbidlist):
        driver = webdriver.Chrome(service=s)
        driver.get("https://biobank.ndph.ox.ac.uk/ukb/field.cgi?id=" + str(ukbbid))
    # Search by XPATH (put XPATH between """___""")
        Description.append(driver.find_element(by=By.XPATH, value="""//*[@id="main"]/table[1]/tbody/tr[1]/td[2]""").text)
        Category.append(driver.find_element(by=By.XPATH, value="""//*[@id="main"]/table[1]/tbody/tr[2]/td[2]/a[1]""").text)
        ValueType.append(driver.find_element(by=By.XPATH, value="""//*[@id="main"]/table[2]/tbody/tr[1]/td[5]""").text)
        Nbpart.append(int(driver.find_element(by=By.XPATH, value="""//*[@id="main"]/table[2]/tbody/tr[1]/td[2]""").text.replace(",", "")))
        Sex.append(driver.find_element(by=By.XPATH, value="""//*[@id="main"]/table[2]/tbody/tr[1]/td[8]""").text)
    # Search by text element (write search query e.g. "Mean = " and remove by replacing string)
        try:
            Mean.append(driver.find_element(By.XPATH, value = ("//*[contains(text(),'Mean = ')]")).text.replace("Mean = ", ""))
        except NoSuchElementException:
            Mean.append("NA")
    
        try:    
            Median.append(driver.find_element(By.XPATH, value = ("//*[contains(text(),'Median')]/following::td")).text)
        except NoSuchElementException:
            Median.append("NA")

        try:
            Stdv.append(driver.find_element(By.XPATH, value = ("//*[contains(text(),'Std.dev = ')]")).text.replace("Std.dev = ", ""))
        except NoSuchElementException:
            Stdv.append("NA")

        driver.close()
    # Write dataframe
    df = pd.DataFrame({'FieldID':ukbbidlist,'Description':Description, "Category":Category,
                       "Value_Type":ValueType, "Nb_Participants": Nbpart,
                       'Sex':Sex,"Mean":Mean, "Median": Median, "SD":Stdv}) 
    
    return(df)

### From TL_metadata file

### Load dataframe
df_metadata = pd.read_csv(r'TL_metadata.csv')
### Get list of FieldIDs
fieldids = list(df_metadata["FieldID"])
fieldids = [x for x in fieldids if str(x) != 'nan'] # Remove NaN
fieldids = [int(i) for i in fieldids] # Convert float to int

### Extract information
df = UKBB_to_DF(fieldids)

df

df.to_csv('pheno_list_f.csv', index=False, encoding='utf-8')
df.to_excel('pheno_list_f.xlsx', index=False, encoding='utf-8')
